+++ src/python/dxpy/bindings/dxproject.py
@@ -36,6 +36,7 @@ from __future__ import print_function, unicode_literals, division, absolute_impo
 
 import dxpy
 from . import DXObject
+from ..exceptions import DXError
 
 ###############
 # DXContainer #
@@ -201,10 +202,15 @@ class DXContainer(DXObject):
         if isinstance(self, DXProject):
             api_method = dxpy.api.project_remove_folder
 
-        api_method(self._dxid,
-                   {"folder": folder, "recurse": recurse, "force": force},
-                   always_retry=force,  # api call is idempotent under 'force' semantics
-                   **kwargs)
+        completed = False
+        while not completed:
+            resp = api_method(self._dxid,
+                              {"folder": folder, "recurse": recurse, "force": force, "partial": True},
+                              always_retry=force,  # api call is idempotent under 'force' semantics
+                              **kwargs)
+            if 'completed' not in resp:
+                raise DXError('Error removing folder')
+            completed = resp['completed']
 
     def remove_objects(self, objects, force=False, **kwargs):
         """